Added InitPeerIsn to retrieve ISNs. Useful for MPTCP relative mappings. Next commit will serialize DSS options with relative sequence numbers instead of the current ones

+++ src/internet/model/tcp-rx-buffer.cc
@@ -127,7 +127,7 @@ TcpRxBuffer::IncNextRxSequence ()
   NS_LOG_FUNCTION (this);
   // Increment nextRxSeq is valid only if we don't have any data buffered,
   // this is supposed to be called only during the three-way handshake
-  NS_ASSERT (m_size == 0);
+  NS_ASSERT (m_size == (uint32_t)0);
   m_nextRxSeq++;
 }
